update action, add error-handling

+++ src/Extract-PBIXMetadata.ps1
@@ -3,7 +3,6 @@ $ErrorActionPreference = "Stop"
 # print Information stream
 $InformationPreference = "Continue"
 
-
 $root_path = Switch ($Host.name) {
 	'Visual Studio Code Host' { split-path $psEditor.GetEditorContext().CurrentFile.Path }
 	'Windows PowerShell ISE Host' { Split-Path -Path $psISE.CurrentFile.FullPath }
@@ -24,6 +23,7 @@ $git_event_after = $env:GIT_EVENT_AFTER
 $workspace_id = $env:PBI_PREMIUM_WORKSPACE_ID
 
 if((Test-Path 'env:PBI_TENANT_ID') -and (Test-Path 'env:PBI_CLIENT_SECRET') -and (Test-Path 'env:PBI_CLIENT_ID')) {
+	Write-Information "Using Service Principal authentication!"
 	$tenant_id = $env:PBI_TENANT_ID
 	$client_id = $env:PBI_CLIENT_ID
 	$client_secret = $env:PBI_CLIENT_SECRET
@@ -35,6 +35,7 @@ if((Test-Path 'env:PBI_TENANT_ID') -and (Test-Path 'env:PBI_CLIENT_SECRET') -and
 	Connect-PowerBIServiceAccount -Credential $credential -ServicePrincipal -TenantId $tenant_id
 }
 else {
+	Write-Information "Using Username/Password authentication!"
 	$user_name = $env:PBI_USER_NAME
 	$user_password = $env:PBI_USER_PASSWORD
 	$login_info = "User ID=$user_name;Password=$user_password"
@@ -58,32 +59,49 @@ Write-Information $pbix_files
 
 foreach($pbix_file in $pbix_files)
 {
-	Write-Information "Processing  $($pbix_file.FullName) ... "
-	$temp_name = "$($pbix_file.BaseName)-$(Get-Date -Format 'yyyyMMddTHHmmss')"
-	Write-Information "Uploading $($pbix_file.FullName) to $($workspace.Name)/$temp_name ... "
-	$report = New-PowerBIReport -Path $pbix_file.FullName -Name $temp_name -WorkspaceId $workspace.Id
-	Start-Sleep -Seconds 5
-	Write-Information "    Done!"
+	$report = $null
+	$dataset = $null
+	try {
+		Write-Information "Processing  $($pbix_file.FullName) ... "
+		$temp_name = "$($pbix_file.BaseName)-$(Get-Date -Format 'yyyyMMddTHHmmss')"
+		Write-Information "Uploading $($pbix_file.FullName) to $($workspace.Name)/$temp_name ... "
+		$report = New-PowerBIReport -Path $pbix_file.FullName -Name $temp_name -WorkspaceId $workspace.Id
+		Start-Sleep -Seconds 5
+		Write-Information "    Done!"
 
-	Write-Information "Getting PowerBI dataset ..."
-	$dataset = Get-PowerBIDataset -WorkspaceId $workspace.Id | Where-Object { $_.Name -eq $temp_name}
-	$connection_string = "powerbi://api.powerbi.com/v1.0/myorg/$($workspace.Name);initial catalog=$($dataset.Name)"
+		Write-Information "Getting PowerBI dataset ..."
+		$dataset = Get-PowerBIDataset -WorkspaceId $workspace.Id | Where-Object { $_.Name -eq $temp_name}
+		$connection_string = "powerbi://api.powerbi.com/v1.0/myorg/$($workspace.Name);initial catalog=$($dataset.Name)"
 
-	$executable = Join-Path $root_path tools TabularEditor2 TabularEditor.exe
+		$executable = Join-Path $root_path TabularEditor.exe
 
-	$params = @(
-		"""Provider=MSOLAP;Data Source=$connection_string;$login_info"""
-		"""$($dataset.Name)"""
-		"-BIM $(Join-Path $pbix_file.DirectoryName $pbix_file.BaseName).database.json" 
-	)
+		$params = @(
+			"""Provider=MSOLAP;Data Source=$connection_string;$login_info"""
+			"""$($dataset.Name)"""
+			"-BIM $(Join-Path $pbix_file.DirectoryName $pbix_file.BaseName).database.json" 
+		)
 
-	Write-Information "$executable $params"
-	$p = Start-Process -FilePath $executable -Wait -NoNewWindow -PassThru -ArgumentList $params
+		Write-Information "$executable $params"
+		$p = Start-Process -FilePath $executable -Wait -NoNewWindow -PassThru -ArgumentList $params
 
-	Write-Information $p.ExitCode
-
-	Write-Information "Removing temporary PowerBI report ..."
-	Remove-PowerBIReport -WorkspaceId $workspace.Id -Id $report.Id
-	Write-Information "Removing temporary PowerBI dataset ..."
-	Invoke-PowerBIRestMethod -Url "https://api.powerbi.com/v1.0/myorg/groups/$($workspace.Id)/datasets/$($dataset.Id)" -Method Delete
+		if($p.ExitCode -ne 0)
+		{
+			Write-Error "Failed to extract .bim file from $($dataset.WebUrl)!"
+		}
+	}
+	catch {
+		
+	}
+	finally {
+		if($report -ne $null)
+		{
+			Write-Information "Removing temporary PowerBI report ..."
+			Remove-PowerBIReport -WorkspaceId $workspace.Id -Id $report.Id
+		}
+		if($dataset -ne $null)
+		{
+			Write-Information "Removing temporary PowerBI dataset ..."
+			Invoke-PowerBIRestMethod -Url "https://api.powerbi.com/v1.0/myorg/groups/$($workspace.Id)/datasets/$($dataset.Id)" -Method Delete
+		}
+	}
 }
